pyproject.toml
[project]
name = "wayback-pdf-diff"
version = "0.1.0"
authors = [
    {name = "Internet Archive"},
]
maintainers = [
    {name = "Internet Archive"},
]
description = "Tools for diffing & comparing PDF content, producing output compatible with web-monitoring-diff."
requires-python = ">=3.10"
license = "AGPL-3.0-only"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Environment :: Console",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Archiving",
]
dependencies = [
    "pymupdf>=1.24.0",
    "fast-diff-match-patch>=2.0.1,<3",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
server = [
    "tornado>=6.0",
]

[project.scripts]
wayback-pdf-diff = "wayback_pdf_diff.server:cli"

[project.urls]
Homepage = "https://github.com/internetarchive/wayback-pdf-diff"
Issues = "https://github.com/internetarchive/wayback-pdf-diff/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["wayback_pdf_diff", "wayback_pdf_diff.server"]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
